Add files via upload
SAS Code to get data for model. To start first need to take all "combined_data" 1-4 in the Kaggle link and replace every ":" with ", ,            "
Also grab the movie titles data from the kaggle (more for interest, we didn't have time to search titles for underlying movie information that could have been useful)
This is since the dataset is set up weirdly with movies first then user, rating, date in an uneven order behind it. Since it is only 4 "control H commands felt easier than doing more work for formatting input code. Then once all code is in the same folder everything else is set in the work environment to not clog up disk space. Follow order and it will run but after a pretty long time.